refactor: created base test case as module
