+++ tests/test_authentication.py
@@ -1,4 +1,4 @@
-from tests import ApiTestCase
+from tests.base_test_case import ApiTestCase
 from app.models import User
